// packages/parser-sdk/src/types/ParserCapabilities.ts
export interface ParserCapabilities {
  supportsClasses: boolean;
  supportsInterfaces: boolean;
  supportsEnums: boolean;
  supportsNamespaces: boolean;
  supportsImports: boolean;
  supportsExports: boolean;
  supportsDecorators: boolean;
  supportsGenerics: boolean;
  supportsAnnotations: boolean;
  supportsMacros: boolean;
}